# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "x0tta6bl4 Mesh Network API",
    "version": "3.2.1",
    "description": "\n# x0tta6bl4 Mesh Network API\n\nSelf-healing mesh network node with MAPE-K autonomic loop and Kimi K2.5 Agent Swarm.\n\n## Features\n\n- **Post-Quantum Cryptography**: ML-KEM-768, ML-DSA-65 (NIST FIPS 203/204)\n- **SPIFFE/SPIRE Integration**: Zero Trust workload identity with mTLS\n- **Agent Swarm**: Parallel task execution with up to 100 agents\n- **PARL**: Parallel Agent Reinforcement Learning for 4.5x speedup\n- **Vision Module**: Visual analysis for mesh topology and anomaly detection\n\n## Authentication\n\nMost endpoints require `X-Admin-Token` header for authentication.\n\n## Rate Limiting\n\n- Standard endpoints: 100 requests/minute\n- Task submission: 1000 requests/minute\n- Swarm creation: 10 requests/minute\n- Scaling: 5 requests/minute\n\n## Error Handling\n\nAll errors follow RFC 7807 Problem Details format:\n```json\n{\n  \"type\": \"https://x0tta6bl4.mesh/errors/swarm-not-found\",\n  \"title\": \"Swarm Not Found\",\n  \"status\": 404,\n  \"detail\": \"Swarm with ID 'abc123' does not exist\"\n}\n```\n        ",
    "contact": {
      "name": "x0tta6bl4 Support",
      "email": "support@x0tta6bl4.mesh",
      "url": "https://x0tta6bl4.mesh"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "x-logo": {
      "url": "/logo.svg",
      "altText": "x0tta6bl4 Logo"
    }
  },
  "paths": {
    "/vpn/config": {
      "get": {
        "tags": [
          "vpn"
        ],
        "summary": "Get Vpn Config",
        "description": "Generate VPN configuration for a user.\n\nArgs:\n    user_id: Telegram user ID\n    username: Optional username for the VPN config\n    server: Optional custom server address\n    port: Optional custom server port\n\nReturns:\n    VPN configuration with VLESS link and detailed instructions",
        "operationId": "get_vpn_config_vpn_config_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          },
          {
            "name": "username",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Username"
            }
          },
          {
            "name": "server",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Server"
            }
          },
          {
            "name": "port",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Port"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VPNConfigResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "vpn"
        ],
        "summary": "Create Vpn Config",
        "description": "Create VPN configuration for a user.\n\nArgs:\n    request: VPN config request data\n\nReturns:\n    VPN configuration with VLESS link and detailed instructions",
        "operationId": "create_vpn_config_vpn_config_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VPNConfigRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VPNConfigResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/vpn/users": {
      "get": {
        "tags": [
          "vpn"
        ],
        "summary": "Get Vpn Users",
        "description": "Get list of active VPN users.\n\nReturns:\n    List of active VPN users (cached for 60 seconds)",
        "operationId": "get_vpn_users_vpn_users_get",
        "parameters": [
          {
            "name": "x-admin-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Vpn Users Vpn Users Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/vpn/user/{user_id}": {
      "delete": {
        "tags": [
          "vpn"
        ],
        "summary": "Delete Vpn User",
        "description": "Delete VPN user.\n\nArgs:\n    user_id: User ID to delete\n\nReturns:\n    Confirmation of deletion",
        "operationId": "delete_vpn_user_vpn_user__user_id__delete",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          },
          {
            "name": "x-admin-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Delete Vpn User Vpn User  User Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/users/": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get Users",
        "description": "Get all users (admin only)",
        "operationId": "get_users_api_v1_users__get",
        "parameters": [
          {
            "name": "x-admin-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserResponse"
                  },
                  "title": "Response Get Users Api V1 Users  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UserResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "company": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company"
          },
          "plan": {
            "type": "string",
            "title": "Plan"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "full_name",
          "company",
          "plan",
          "created_at"
        ],
        "title": "UserResponse"
      },
      "VPNConfigRequest": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          },
          "server": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Server"
          },
          "port": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Port"
          }
        },
        "type": "object",
        "required": [
          "user_id"
        ],
        "title": "VPNConfigRequest"
      },
      "VPNConfigResponse": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          },
          "vless_link": {
            "type": "string",
            "title": "Vless Link"
          },
          "config_text": {
            "type": "string",
            "title": "Config Text"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "username",
          "vless_link",
          "config_text"
        ],
        "title": "VPNConfigResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "AdminToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Admin-Token",
        "description": "Admin authentication token"
      },
      "mTLS": {
        "type": "mutualTLS",
        "description": "Mutual TLS authentication with SPIFFE SVID"
      }
    },
...
  },
  "servers": [
...
}